adding admin controllers

+++ server/controllers/order.controller.js
@@ -7,7 +7,7 @@ import ApiResponse from "../utils/ApiResponse.util.js";
 
 
 //create order
-export const createOrder = asyncHandler(async (req, res, next) => {
+export const createOrder = asyncHandler(async (req, res) => {
   const { products, shippingAddress } = req.body;
   const userId = req.user._id; // assuming you attach user from auth middleware
 
@@ -39,8 +39,8 @@ export const createOrder = asyncHandler(async (req, res, next) => {
 
 
 
-//get all orders
-export const getMyOrders = asyncHandler(async (req, res) => {
+//get all user orders
+export const getMyOrders = asyncHandler(async (req, res) => {  
   const orders = await Order.find({ user: req.user._id }).populate("products.product");
 
   res
@@ -50,49 +50,23 @@ export const getMyOrders = asyncHandler(async (req, res) => {
 
 
 
-//get one specific order
-export const getOrderById = asyncHandler(async (req, res) => {
-  const order = await Order.findById(req.params.id)
-    .populate("user", "name email")
-    .populate("products.product");
-
-  if (!order) throw new ApiError(404, "Order not found");
-
-  // Only allow the owner or admin
-  if (order.user._id.toString() !== req.user._id.toString() && req.user.role !== "admin") {
-    throw new ApiError(403, "Not authorized to view this order");
-  }
-
-  res.status(200).json(new ApiResponse(200, order, "Order fetched successfully"));
-});
-
-
-
-//update order status
-export const updateOrderStatus = asyncHandler(async (req, res) => {
-  const { status } = req.body;
-  const order = await Order.findById(req.params.id);
-
-  if (!order) throw new ApiError(404, "Order not found");
-
-  order.orderStatus = status || order.orderStatus;
-  await order.save();
-
-  res
-    .status(200)
-    .json(new ApiResponse(200, order, "Order status updated successfully"));
-});
-
-
-//delete order
+//delete order(user only)
 export const deleteOrder = asyncHandler(async (req, res) => {
   const order = await Order.findById(req.params.id);
 
   if (!order) throw new ApiError(404, "Order not found");
 
+  if(order.user.toString() !== req.user._id.toString()){
+    throw new ApiError(403, "Not authorized to delete this order");
+  }
+
+  if(order.orderStatus !== "pending"){
+    throw new ApiError(400, "Only pending orders can be cancelled");
+  }
+
   await order.deleteOne();
 
   res
     .status(200)
-    .json(new ApiResponse(200, null, "Order deleted successfully"));
+    .json(new ApiResponse(200, deleteOrder, "Order deleted successfully"));
 });
